Move CLI to its own crate
