+++ Cargo.toml
@@ -1,7 +1,6 @@
 [workspace]
 resolver = "2"
-# members = ["anise", "anise-gui", "anise-cli"]
-members = ["anise"]
+members = ["anise", "anise-cli"]
 
 [workspace.package]
 version = "0.1.0"
@@ -26,14 +25,11 @@ exclude = [
     "*.sh",
 ]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [workspace.dependencies]
 hifitime = "3.8.6"
 memmap2 = "=0.9.3"
 crc32fast = "=1.3.2"
 der = { version = "0.7.8", features = ["derive", "alloc", "real"] }
-clap = { version = "4", features = ["derive"] }
 log = "=0.4"
 pretty_env_logger = "=0.5"
 tabled = "=0.15"
@@ -46,6 +42,8 @@ snafu = { version = "0.7.5", features = ["backtrace"] }
 lexical-core = "0.8.5"
 heapless = "0.8.0"
 rstest = "0.18.2"
+
+anise = { version = "0.1.0", path = "anise", default-features = false }
 # eframe = { version = "0.24.0", optional = true }
 # egui = { version = "0.24.0", optional = true }
 # egui_extras = { version = "0.24.0", features = [
