+++ anise-cli/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "anise-cline"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+description = "A command line interface for ANISE"
+
+[dependencies]
+anise = { workspace = true }
+clap = { version = "4", features = ["derive"] }
+pretty_env_logger = { workspace = true }
+tabled = { workspace = true }
+bytes = { workspace = true }
+memmap2 = { workspace = true }
+snafu = { workspace = true }
+log = { workspace = true }
+zerocopy = { workspace = true }
+hifitime = { workspace = true }
+
+# eframe = { version = "0.24.0", optional = true }
+# egui = { version = "0.24.0", optional = true }
+# egui_extras = { version = "0.24.0", features = [
+#     "datepicker",
+#     "http",
+#     "image",
+# ], optional = true }
+# egui-toast = { version = "0.10.0", optional = true }
+# rfd = { version = "0.12.1", optional = true }
+
+# [target.'cfg(target_arch = "wasm32")'.dependencies]
+# wasm-bindgen-futures = "0.4"
+# poll-promise = { version = "0.3.0", features = ["web"] }
+
+[[bin]]
+name = "anise-cli"
+path = "src/main.rs"
+
+[features]
