+++ anise-cli/src/inspect.rs
